Migrated coinbase account to the account api
