+++ src/client/RESTClient.ts
@@ -12,7 +12,6 @@ import axiosRetry, {isNetworkOrIdempotentRequestError} from 'axios-retry';
 import util from 'util';
 import {EventEmitter} from 'events';
 import {getErrorMessage, gotRateLimited, inAirPlaneMode} from '../error/ErrorUtil';
-import {CoinbaseAccountAPI} from '../coinbase/CoinbaseAccountAPI';
 
 export interface RESTClient {
   on(
@@ -40,7 +39,6 @@ export class RESTClient extends EventEmitter {
   readonly product: ProductAPI;
   readonly profile: ProfileAPI;
   readonly user: UserAPI;
-  readonly coinbase: CoinbaseAccountAPI;
 
   private readonly httpClient: AxiosInstance;
   private readonly logger: (msg: string, ...param: any[]) => void;
@@ -103,7 +101,6 @@ export class RESTClient extends EventEmitter {
     this.product = new ProductAPI(this.httpClient, this);
     this.profile = new ProfileAPI(this.httpClient);
     this.user = new UserAPI(this.httpClient);
-    this.coinbase = new CoinbaseAccountAPI(this.httpClient);
   }
 
   static stringifyPayload(config: AxiosRequestConfig): string {
